my_web_socket/requirements.txt
websockets
pytest-asyncio